Added vulnerability_assessment functionality and updated GUI
Alerts are checked against a list of past alerts so that duplicates are not presented. Start button can no longer be clicked multiple times, starting multiple instances of sniffing and alert detection. vulnerability_assessment combines network scanner and port scanner to find all hosts and open ports on the network. Does cause lots of traffic, however, so is not currently implemented on NIDS